requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.16.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3